new file for webdictionary services
