+++ src/services/languages.ts
@@ -3,8 +3,33 @@ import dbQuery from '../model/db-query';
 import {
   Language, LanguageDB, convertLanguageTypes,
   KnownLanguageDB, KnownLanguage, convertKnownLanguageTypes,
-  WebdictionaryDB, Webdictionary, convertWebdictionaryTypes,
 } from '../types';
+
+
+const addNew = async function(languageData: Language): Promise<void> {
+  const {
+    id,
+    name,
+    googleTranslateURL,
+    eachCharIsWord,
+    isRightToLeft,
+  } = languageData;
+
+  const ADD_LANGUAGE: string = `
+    INSERT INTO languages 
+    (id, name, google_translate_url, each_word_is_char, is_right_to_left)
+    VALUES 
+    (%L, %L, %L, %L, %L)`;
+
+  await dbQuery(
+    ADD_LANGUAGE,
+    id,
+    name,
+    googleTranslateURL || null,
+    eachCharIsWord,
+    isRightToLeft,
+  );
+};
 
 
 const getAll = async function(): Promise<Array<Language> | null> {
@@ -80,85 +105,11 @@ const addPair = async function(sourceLanguageId: string, targetLanguageId: strin
 };
 
 
-const addWebdictionary = async function(languagepairId: number, name: string, url: string): Promise<void> {
-  const ADD_WEBDICTIONARY: string = `
-    INSERT INTO webdictionaries
-    (language_pair_id, "name", "url")
-    VALUES
-    (%s, %L, %L)`;
-
-  await dbQuery(
-    ADD_WEBDICTIONARY,
-    languagepairId,
-    name,
-    url,
-  );
-};
-
-
-const getWebdictsByPair = async function getWebdictionariesByLanguagePair(languagepairId: number): Promise<Array<Webdictionary> | null> {
-  const DICTIONARIES_BY_PAIR: string = `
-    SELECT * FROM webdictionaries
-    WHERE language_pair_id = %s`;
-
-  const result = await dbQuery(DICTIONARIES_BY_PAIR, languagepairId);
-  if (result.rows.length === 0) {
-    return null;
-  }
-
-  return result.rows.map((dbItem: WebdictionaryDB): Webdictionary => convertWebdictionaryTypes(dbItem));
-};
-
-
-const getPrefWebdicts = async function getPreferredWebdictionaries(userId: number): Promise<Array<Webdictionary> | null> {
-  const PREFERRED_WEBDICTIONARIES: string = `
-    SELECT wd.id, language_pair_id, source_language_id, target_language_id, name, url FROM webdictionaries AS wd
-    JOIN languagepairs AS lp ON wd.language_pair_id = lp.id
-    `;
-
-  const result = await dbQuery(PREFERRED_WEBDICTIONARIES, userId);
-  if (result.rows.length === 0) {
-    return null;
-  }
-
-  return result.rows.map((dbItem: WebdictionaryDB): Webdictionary => convertWebdictionaryTypes(dbItem));
-};
-
-
-const addNew = async function(languageData: Language): Promise<void> {
-  const {
-    id,
-    name,
-    googleTranslateURL,
-    eachCharIsWord,
-    isRightToLeft,
-  } = languageData;
-
-  const ADD_LANGUAGE: string = `
-    INSERT INTO languages 
-    (id, name, google_translate_url, each_word_is_char, is_right_to_left)
-    VALUES 
-    (%L, %L, %L, %L, %L)`;
-
-  await dbQuery(
-    ADD_LANGUAGE,
-    id,
-    name,
-    googleTranslateURL || null,
-    eachCharIsWord,
-    isRightToLeft,
-  );
-};
-
-
 export default {
+  addNew,
   getAll,
   getById,
   getKnownByUser,
-  addNew,
-  addPair,
-  addWebdictionary,
   getStudiedByUser,
-  getPrefWebdicts,
-  getWebdictsByPair,
+  addPair,
 };
